requirements.txt
# cosmic-replay v2 - Python Dependencies

# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pyyaml>=6.0
requests>=2.31.0
urllib3>=2.0.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0